feat(app): Overhaul OT-2 calibration experience
Enable the previously under-development flows to support the new
calibration procedures for the OT-2.

The new development flows are based on a new session model, with a
limited number of endpoints that accept structured commands. The
middleware for this lives in app/src/sessions.

In addition, much more data about calibration is exposed via HTTP in
app/src/calibrations.

The new deck calibration doesn't have much that is different, but the
new pipette offset calibration and tip length calibration flows are
integrated in many different parts of the app ecosystem:
- In the attach pipette flow, you now calibrate pipette offset (and
possibly tip length, if necessary) in the attach flow
- There are also buttons for same in the pipettes card
- In the pre protocol flow, you now calibrate tip length for each kind
of tiprack used in your protocol, possibly also recalibrating pipette
offset if you recalibrate a tip length used for pipette offset
- You must have a pipette offset and deck calibration to enter the
protocol flows
- Robot calibration data is now exposed on the robots pages

Closes #6736, #6737