fix(robot-server): Fix return tip moves in cal flows
While we do save the point at which the user indicated a tip was picked
up, and move to that point when we return the tip, _before_ we move to
that point we always were moving to the original tip rack well 0 top +
offset. Then, we'd do a single move to the point indicated by the
user (with the return tip height multiplier applied, so maybe inside the
well). If the user jogged at all, this would be the kind of diagonal XYZ
move that we like to avoid, because particularly if the destination
point is inside the well - because of return tip height - the tip can
hit the rack geometry on the way to its destination.

Instead, refactor the logic a bit to simplify it; make the user flow
utils only interact with public functions; and change the
move-to-tiprack command to move to the saved point if available. This
should mean that the move into the rack to return the tip is always a
pure-vertical move.