fix(app): put blocking user flows in new top level modal portal
There are a number of new calibration calibration flows that launch a wizard style modal in response
to creating a calibration session on the robot. These modals were, like all existing modals in the
app, rendered into a portal component that is a sibling of the two side panel components in the main
layout component. This meant that users were able to interact with the navigation tabs and the
contents of the sidepanel during these captive flows that involve precise coordination and movement
on the robot. In order protect against uncharacterized behavior when users navigate away from the
the current tab during a calibration user flow, this adds a new top level modal portal. Modals
rendered in this portal will span the entire window of the app, and provide assurance that a session
will be deleted any time a user navigates away from, or completes a calibration user flow.