fix(robot-server): home plungers in cal flows
Cal flows in general don't home before hand. This is a reasonably
defensible position since we home when we boot and in general we home a
lot. But what we don't home when we boot is plungers, and so we now home
the active plunger in each cal flow. This fixes an issue where the first
time you ran a cal flow, when going to pick up the tip the tip ejector
shroud might interfere with the pickup.