refactor(app): add intents and specify cal intro
We have a variety of different calibration session types, which cover
the multiple different sequences of actions that the app and server go
through, and the different kinds of data that the server saves. But
there's another layer that differentiates calibration actions from each
other: the intent of the user, driving the context in which they begin a
calibration process.

The intent is a client-side thing. It might be different for each
client (although we only have one). There are often several different
intents with which a calibration process can be started - and the server
doesn't, and shouldn't have to, care about any of them. The only thing
that matters to the server is the actions it has to perform, and the
intent doesn't matter there.

Where the intent does matter, and where the new values are used in the
app, is in explaining to the user what's going on. There's a couple
different calibration flows that can be launched from several different
places based on different user intents, and the user wants to know, for
instance, "Why is this thing talking about tip length calibration when I
clicked calibrate pipette offset?" or "Why is this thing talking about
pipette offset when I clicked recalibrate tip"? And this can be in the
same exact calibration process.

Using intents, we can and now do display different content to the user
depending on their intent when they clicked the button, as expressed by
the context in which the button click occurred.