feat(robot-server): cal: actions to home and retry
Sometimes, when a user is jogging the robot, they crash. Right now,
they're up a creek - all they can do is quit and restart. But what we
can do to give them options is provide actions in states that allow
jogging to home, drop the tip (if there is one), and go back to pick
up another tip (if necessary) and try again.