refactor(app): Add crash handling components
The ConfirmCrashRecoveryModal is a modal from our designs that will be
used to ask users to confirm whether they want to recover from a crash,
with appropriate content changes based on whether they need to replace a
tip in a tiprack or not.

The useConfirmCrashRecovery hook is a higher level interface to that
modal, that both provides the content that is used to invoke the modal
and handles the dispatch of the actual action itself.