feat(app): Add recover from crash buttons to cal
In any calibration panel that involves jogging, we now display a link at
the bottom that allows the user to try and recover if they jogged too
far. If the pipette has a tip, it will be dropped (they should put
another tip in A1); if not, it won't be. In either case it homes.