feat(app,robot-server): Retry cal actions (#6830)
Sometimes, when a user is jogging the robot, they crash. Right now,
they're up a creek - all they can do is quit and restart. But what we
can do to give them options is provide actions in states that allow
jogging to home, drop the tip (if there is one), and go back to pick
up another tip (if necessary) and try again.

In any calibration experience that involves jogging, there is now a little link
at the bottom. If you click it it pops up a ConfirmCrashRecoveryModal (a
modal from our designs that will be used to ask users to confirm whether
they want to recover from a crash, with appropriate content changes 
based on whether they need to replace a tip in a tiprack or not). This is
integrated on a technical level using the useConfirmCrashRecovery hook,
a higher level interface to that modal, that both provides the content that
is used to invoke the modal and handles the dispatch of the actual action itself.

Closes #6729