feat(components): Add StackedLabeledControl
This is a variant of LabeledControl where instead of laying out the
label and children beside the control, the label and children are
layed out full-width above the control. This ends up being a pretty big
difference in layout because the order in which some things are rendered
has to change, so, new control. Takes advantage of now being able to
override the ControlInfo className.