refactor(app): Always display cal block selection
We previously were displaying a toggle for use calibration block which
would allow the user to set it to true or false; since this value is
actually nullable (signifying unset), we only showed it if it had a
value.

Instead, we want to always display a place to change the value,
including back to unset/always-prompt. This is critical for some changes
we're making to physical inclusions in the packaging of the calibration
block - we want to point users to a thing to change in the app, and it
would be annoying if that wasn't always there.

Since this is a tristate setting, we use the new LabeledRadioGroup
control to allow selection of block/trash/unset.

Closes #6872