fix(api): recache position before plunger home
When you're homing the plunger through its top-level home attribute,
there are some particularities around the way the position cache is
updated that can leave you with a slightly-different position cache than
what you started with. A big part of the reason why is, well, floating
point math + inverting matrices.

If the difference is big enough, then a move right after a home plunger
- for instance, the move of the plunger back to its bottom position -
will also imply a move in other axes like X and Y.

And finally, if you do that - have a normal move on a plunger axis, and
a really tiny move on X or Y or both - then the plunger stalls for some
reason.

This is not quite the root cause of #6879 - that lies deeper in
either something the smoothie is doing in acceleration planning, or
something the power supply is doing - and it isn't quite the right fix
for #6879 - that involves fixing the stupid reverse calculations and
making home_plunger not perturb axes other than the ones that it's
homing - but it is pretty close to both of them.

Closes #6879