fix(build): Dont break make with weird deps in shared-data (#6857)
Calling find to look for json dependencies in the shared-data python
makefile was picking up the ones in the build directory, and is a bit
gross anyway. By not doing that anymore (in favor of writing them in
with wildcards) we get the same behavior but more reliably, and now
building wheels and thus deploying to pypi should work a bit better.