refactor(robot-server): Home when initializing calibration sessions
Now that we display UI when a session is being established, it's
reasonable to do something that takes a while like homing before doing
calibration. This also handles a couple other issues:
- You don't have to home the plunger before picking up tip anymore;
since we home everything before hand, the problem is solved
- Since the home happens during session establish, by the time you get
toe the deck setup phase the head is out of the way.