refactor(app): Re-layout pipettes in calibration card
This matches the new designs.

The per-pipette information is set out in a tinted-background block (a
new concept, but necessary with this level of hierarchical information)
and what is used as a title vs as content has changed a little bit to
really put tip length and offset being different forward.