docs(api): cache v1 docs sources and change banner
Rather than capturing the output of the docs build like we did for ot1,
make some def-and-docstring-only copies of the apiv1 sources required
for the v1 api docs (only the stuff that is actually called out in the
automodule) and quit importing the rest of it.

Also, update the banner text to tell people that this is no longer
supported.

Closes #6734