fix(app): Fix intro screen layout with p1k tiprack (#6932)
The p1000 tiprack image is a different size than the other images to
capture the taller tiprack. This means that it would expand the
horizontal layout block. What we really want is pretty fixed-size
blocks, and images that are resized to fit if necessary.

By doing some Messing Around With Flexbox, we can achieve this but it
requires setting an explicit height for the image so the max height
stuff actually works because of the lovely and 100% intuitive tidbit
that max height applies to the parent's defined height not its
max-height, and if you don't have an explicit height defined on the
apparent max-height does
nothing (https://www.w3.org/TR/CSS21/visudet.html#min-max-height).