refactor(app): clear all sessions for robot when disconnected, disable cal buttons (#6962)
In order to prevent sessions from sticking around in redux state after a robot has disconnected,
purge all sessions that belong to the disconnected robot from redux state. All buttons on the
pipette info tab should be disabled with a tooltip if not connected to the robot, just as the
buttons on the robot page are.