refactor(app): call out removable plastic trash bin in cal flow intros (#6994)
Highlight that the removable black plastic trash bin is required for Tip Length Calibration,
Extended Pipette Offset Calibration, and Calibration Health Check if the user doesn't have access to
a Calibration Block. In these cases, list the trash bin as a required labware with an image in the
Introduction panels of these flows.