chore(build): workaround set-env deprecation (#7010)
Github recently disabled the ::set-env command in their protocol runners because it was insecure (https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/). The fix for this was to use environment file piping instead. Unfortunately, even if you're interacting with it through e.g. a node wrapper like in https://github.com/actions/core, this is what you're always doing, so core needed a bump to v1.2.6 in all its dependencies. This didn't happen in https://github.com/actions/github-script as of this writing (open pr: actions/github-script#91) and therefore our builds started breaking.

The workaround is to enable the old behavior and have a message that yells at you about it.