fix(app): prompt whether to replace pipette after detach
The addition of a loading button state to the confirm pipette is detached button in the change
pipette flow, silently broke the wizard continuing onto the page that prompts the user to decide
whether or not they want to replace the pipette. Because that request tracking logic lives in a hook
on the check pipettes button component, lets make sure that it doesn't unmount, and just hides, to
avoid accidentally skipping the onDone callback that it receives as a prop.