refactor(api, robot-server): per-revision configs
With the launch of the OT-2 refresh, we'll need to be able to express
some config elements (specifically, currents) differently based on the
hardware revision of the robot in question. This PR adds per-revision
concepts to the robot configs area, specifically around the currents,
which is a huge changeset.

It also has a bunch of refactors that were waiting a long time, like
- robot configs is a dataclass now
- A lot had to deal with that
- Removed the rest of the now-unused config elements.

Closes #7093