feat(robot-server): expose uvicorn logs
We currently expose nice ways to download logs from the opentrons api
package, and from the serial port handlers. What that leaves as a hole
is logging from the server, which encompasses both the built in response
logs (very useful for figuring out interaction patterns) and errors that
bubble up to the server logging handlers.

By adding a URI to the health response logs array, we can get the app to
download the server logs quite easily.

Closes #7097