refactor(app): manage copy strings via i18next in robot settings page (#7119)
Introduce the internationalization library i18next (and adjacent libs for our workflow) to manage
the copy strings in the app. Organize translations into namespaces that encapsulate areas of
functionality within the app (e.g. 'robot_calibration', 'robot_connection'), as well as a 'shared'
namespace for copy strings that are used across the app. Migrate all copy within the RobotSettings
page and its children to make use of i18next via the 'react-i18next' utility hook `useTranslation`,
and component `Trans` as a model of how this migration process can be adopted across the rest of the
run app. This provides a much cleaner solution for copy string management, that is aware of html
element interpolation and context variants like pluralization by default. This also lays necessary
long-term foundation work for the ability to localize the entire app for various other
languages/locales/cultures.