fix(api): do not call a method dependent on the Python API Version directly
The return tip property was added in 2.2 of the python api. If you are on an API version earlier than 2.2, and you try to calibrate your labware in subsequent versions
you will get an error and the robot will appear to not be able to move forward in labware calibration.