refactor(app): reorganize redux modules into their own directory (#7283)
Create a sub directory to contain all redux specific logic within the app/src/ directory. As this is
a first step to considering the modularity of run app's many chunks of functionality, this also adds
a dev script to help visualize the connections between different react sub trees from the
app/src/components directory and different redux modules. The next step is to turn an eye to the
organization of our react components to make this visualization clearer.