fix(api): fix mount critical point (#7314)
We used to consider the p300 gen 1 to be 0mm long. That means that the
smoothie coordinate system was set so 0 is with the end of such a
pipette touching a deck, and that's what the mount critical point was
set to work with.

We don't do that anymore. That means that _now_, our mount critical
point needs to have a -30mm offset unless we want to change the home
positions on the motor controller, which we very much don't.

Closes #7092