refactor(app): reorganize view layer files with atomic principles (#7358)
This organizes the react components that were split between the components and pages directories
into more meaningful top level directories. atoms contains all react components which are so so
narrow in scope that to break them down further wouldn't make sense in code or when rendered.
molecules contains all react components which are reusable potentially and constructed of many atoms
and other use case specific bits, but receive their external state only through props. organisms
contains all react components which a large and tightly bound to a single feature, but may not be
bound to a single location within the app nav hierarchy, they may also hook into state outside of
the props provided to them. the pages directory contains app specific components and still contains
a top level mapping of the global nav in the app, it contains a sub dir for every tab in the App,
within which components which are tightly bound to that navigational slice are collocated within sub
directories.