Liquid sense probing
In order to get feedback from the robot, I soldered a gpio pin to the deck probe pin to trigger the liquid probe sensing. This was to allow us to perform faster liquid sensing. In a test script, I change the state of the gpio pin 10 to make the probe function stop.