fix(discovery-client): allow discovery client to track network interface changes
Rather than rely on its caller, this change allows the Discovery Client to check the system's
network interfaces to determine if its mDNS browser needs to be restarted to bind on any newly added
interfaces (or to flush out any removed interfaces). This change is necessary to increase Discovery
Client reliability when robots connected to U2E adapters are plugged and unplugged from the user's
computer.