feat(api): expand emulators and integration tests. (#7739)
* expand integration tests.

* comments and __eq__ in OptionalValue

* Modify Parser to look for any gcode rather than only a preselected set. This was hiding a bug. If Parser is initialized with G1 as a known gcode and parsed a line like 'G1 G32 F12', the result would be {gcode:G1, params:{G:32, F12}} which is very wrong.
Parser will raise if line starts with non-gcode.

* remove driver_3_0._byte_array_to_hex_string in favor of utils.string_to_hex.
use string to hex to initialize pipette ids and models in smoothie emulator.
in smoothie emulator, homing status is set to true on home commands and is not reset on movement. once homed the homing statuses will remain true. we need to add a simulation of halt pin to reset these flags.

* little fixes.

* more cleanup.

* skip the tempdeck temperature integration test because it sometimes fails in CI. this test will soon not require sleeps.