chore(app-shell): contextIsolation back to false
In electron 12, the contextIsolation feature offered by electron to
prevent several very common vulnerabilities triggerable from
user-controlled content in the renderer was defaulted to true. We should
use it! But for it to do anything useful, the ipc interface between
renderer and main process has to be made much more explicit and exposed
across the isolation layer on a per-method (and with sanitization on the
method inputs) basis, so for now just turn it back off.