chore: upgrade to Electron 13 and Node 14 (#8226)
Upgrade our desktop app stack to electron 13 (from 6) and node 14 (from 12).

Electron 13 has support for a wayland direct renderer without needing to
use x11 bridging which will be quite useful for OT3. It's also a good
idea to catch up on 7 major versions worth of improvements.

The only code changes are to catch up with breaking changes in 7 versions worth of electron.

In electron 12, the contextIsolation feature offered by electron to
prevent several very common vulnerabilities triggerable from
user-controlled content in the renderer was defaulted to true. We should
use it! But for it to do anything useful, the ipc interface between
renderer and main process has to be made much more explicit and exposed
across the isolation layer on a per-method (and with sanitization on the
method inputs) basis, so for now just turn it back off.

https://www.electronjs.org/docs/breaking-changes#removed-shellmoveitemtotrash
https://www.electronjs.org/docs/breaking-changes#api-changed-shellopenitem-is-now-shellopenpath

Co-authored-by: Mike Cousins <mike@opentrons.com>