refactor(api): allow new keywords in serial driver
The serial driver looks for the literals "error" and "alarm" in
responses to indicate conditions that should raise exceptions, since
that's what the smoothie and some modules do. The heater shaker uses ERR
for these cases. Add the ability to specify different alarm and/or error
keywords to the serial connection that will cause the exceptions.