Avoid cross-referencing to `ModuleContext.load_labware()`.
`ModuleContext` is not something that appears in our documentation (nor IMO should it be), so there's nowhere for these cross-references to go.

Instead, link individually to the `.load_labware()` methods of `ThermocyclerContext`, `TemperatureModuleContext`, and `MagneticModuleContext`. Where this would be very clunky, just use the literal text `module.load_labware()` without a link.