Don't cross-reference to `MAX_SUPPORTED_VERSION` or `APIVersion`.
These do not currently appear in our docs, so there's nowhere for these cross-references to go.

Whether `APIVersion` and `MAX_SUPPORTED_VERSION` should be part of our public API, and, if so, to what extent, is a question that's slightly too nontrivial for this PR. For now, I'm shutting up the warnings by turning these mentions into unlinked text, which preserves how they're currently rendered.