chore(api): Add heater/shaker serial driver (#8422)
- refactor(api): allow new keywords in serial driver

The serial driver looks for the literals "error" and "alarm" in
responses to indicate conditions that should raise exceptions, since
that's what the smoothie and some modules do. The heater shaker uses ERR
for these cases. Add the ability to specify different alarm and/or error
keywords to the serial connection that will cause the exceptions.

- feat(api): Add heater-shaker driver

Add a basic heater-shaker serial driver to the API.

- feat(g-code-testing): Add Heater-Shaker G-Codes to G-Code Parser (RET-275) (#8503)

Co-authored-by: Derek Maggio <DerekMaggio@users.noreply.github.com>