chore: add push function for sdists to ot3
While the OT2 pushes work by unzipping an egg into the pythonpath, here
we have separate python paths per app, and a different place to do the
unzips in, which is a lot of differences. Plus, in reality, the actual
build process is working on sdists. While installing wheels in the same
way would work, to match the real build practice we can cut out some
middlemen and just push and unzip sdists to the app directories.