refactor(api,robot-server): sd_notify in server (#8536)
sd_notify() is a systemd api that a program being started in a unit may
call to inform pid1 that it is done starting up. It should be done _by
the program_, based on the program's estimation of when it's ready.

In the implementation where this call happened in the api
subproject (opentrons package), if that package was not initialized -
say, because the robot server wasn't using it, instead relying on a
server-level simulation - then sd_notify() wouldn't be called, and the
robot server application would never be seen as done starting up.

By moving the sd_notify() call to the actual _application_, the problem
is solved, and even if we're not using the subpackage the application
itself will now properly notify that it has come up.

This requires some changes to the python packaging/versioning
infrastructure. Specifically, the systemd_python dependency moves from
the dependencies of the opentrons package to the robot_server package.
This requires removal from the opentrons setup.py and Pipfile, and a
relock there - which requires some fixups to keep package minor versions
to avoid having to change versions on the robot - and addition to the
robot_server setup.py and Pipfile, and a relock there, and some of the
same modifications.