chore(build): allow pushing to ot3s (#8527)
Add makefile targets to push code to toradex verdins as in the OT3.

- fix(shared-data): fix python sdist generation

The sdist here wasn't properly copying data files into the wheel, and
now it is.

- chore: add sdist name getter

This is used to build the name of an sdist tarball in the same way we
have one to build the name of a wheel since it folds in the version.

- chore: add push function for sdists to ot3

While the OT2 pushes work by unzipping an egg into the pythonpath, here
we have separate python paths per app, and a different place to do the
unzips in, which is a lot of differences. Plus, in reality, the actual
build process is working on sdists. While installing wheels in the same
way would work, to match the real build practice we can cut out some
middlemen and just push and unzip sdists to the app directories.

- chore: add push tasks for python to ot3

These new makefile tasks will build up sdists and push them over to the
ot3 in the right places. It works basically the same way as the pushes
for ot2.