feat(api): Add hardware control for heater shaker (#8435)
This adds a hardware controller for the heater-shaker module.

It follows the implementation of other modules and exposes the subset of production APIs for the heater shaker. It does not yet implement firmware upload. It may need further work in the future that becomes apparent once protocol engine begins to actually use it, but for now it can bind to and send commands to heater shaker modules.