chore(api,shared-data): pipenv not for libraries
Pipenv is meant for applications, not libraries, and treating
Pipfiles (and pipenvs)
in library roots as actual sources of truth for dependency versions when
they're not used by

- any robot builder (the dependencies used are those from the setup.py
or pipenvs of the top level applications, not these library dependencies)
- users installing via pip (that's setup.py, what's in the install_requires)

just confuses things.

Instead, we can just use the library's setup.py source of truth as a dev
dependency and Pipenv remains useful for creating environments for tests
and development, without confusing anything.