feat(update-server): add openembedded implementation
The update server's openembedded (OT3) implementation looks strikingly
similar to the buildroot implementation, but relies on a command line
utility called bmap-tools to actually flash images to the hardware.
Bmap-tools is here: https://github.com/intel/bmap-tools . For all that
it looks exactly like a Python library, it is a command line tool that
happens to be written in python and doesn't work as a library, and is
used as such - it's not a python dep and can't be.