fix(api): check home status in move_rel
move_rel (and move_to and current_position) have the problem of checking
if the system is homed by checking if the internal _current_position
dict is truthy. Really all that does is check if the hardware controller
has ever fetched the current position from the smoothie - really nothing
to do with homing.

Homing status is kept in the smoothie driver but hadn't previously been
exposed to the hardware controller.

We can check if the axes that will move during a move call are homed by
exposing that homed-flag data up to the hardware controller and checking
it in place of the is-the-position-known check.

Unfortunately, this breaks a lot of tests! That almost certainly means
the tests are wrong, because if these functions were being called on a
real unhomed robot they'd cause exceptions; but there's a lot of wrong
tests.

For now, gate the new behavior on a default-False optional argument to
specifically move_rel, but we should fix this in the future.