refactor(app): animate expand and collapse for setup accordions
Animate expanding and collapsing the protocol setup accordion steps to guide the user through the
process in a clearer manner. There is now a short delay upon landing on the protocol setup page and
the expansion of the default initial step. This is meant to provide visual context to the user. This
also includes a couple of layout tweaks to the sizing of the labware and module setup step's deck
maps and their surrounding whitespace.

Closes #8471