refactor(update-server): factor out common subpackage (#8611)
The update server will soon support the OT3 as well as the OT2. Most of
the behavior of the update server will remain exactly the same, but some
stuff needs to change - for instance, the OT3 update process will likely
use a tool called bmap that allows for a larger root partition size
while keeping updates only as large as the actual contents of the
filesystem.

This commit moves such shared code to a new subpackage otupdate.common,
which will contain endpoints and things like ssh key management and
update file hash and signature checking, which will stay the same.