chore(hardware): add firmware python module as dep
Pipenv can add git repo dependencies, so we can add the
opentrons_ot3_firmware repo here.

If you want to switch branches you can edit the pipfile (there's not
really a better way to do this).

This will become a problem if we ever want to depend on this module and
get its transitive dependencies because that stuff only reads setup.py,
but we don't really do that yet.